Revert "add a hook to lingua to allow speicifc provider translation after universal translation"
This reverts commit 41136c29b5dddb8448f405ba5a9b12d465a06cc5.
