+++ crates/braintrust-llm-router/src/providers/bedrock.rs
@@ -1,5 +1,3 @@
-use std::future::Future;
-use std::pin::Pin;
 use std::time::{Duration, SystemTime};
 
 use async_trait::async_trait;
@@ -12,8 +10,6 @@ use aws_smithy_runtime_api::client::identity::Identity;
 use bytes::Bytes;
 use http::Request as HttpRequest;
 use lingua::serde_json::Value;
-use lingua::universal::message::{Message, UserContent, UserContentPart};
-use lingua::util::media::MediaBlock;
 use reqwest::header::{HeaderMap, HeaderValue, CONTENT_TYPE};
 use reqwest::Url;
 
@@ -23,100 +19,7 @@ use crate::client::{build_middleware_client, ClientSettings};
 use crate::error::{Error, Result, UpstreamHttpError};
 use crate::providers::ClientHeaders;
 use crate::streaming::{bedrock_event_stream, RawResponseStream};
-use lingua::{ProviderFormat, TransformError};
-
-const BEDROCK_REMOTE_MEDIA_MAX_BYTES: usize = 5 * 1024 * 1024;
-
-type FetchMediaFuture<'a> = Pin<Box<dyn Future<Output = Result<MediaBlock>> + Send + 'a>>;
-
-fn is_remote_image_url(value: &str) -> bool {
-    value.starts_with("http://") || value.starts_with("https://")
-}
-
-async fn fetch_remote_image_as_base64(url: &str) -> Result<MediaBlock> {
-    lingua::util::media::convert_media_to_base64(url, None, Some(BEDROCK_REMOTE_MEDIA_MAX_BYTES))
-        .await
-        .map_err(|e| Error::InvalidRequest(format!("failed to fetch image URL {url}: {e}")))
-}
-
-fn should_inline_remote_image_urls(format: ProviderFormat) -> bool {
-    matches!(
-        format,
-        ProviderFormat::BedrockAnthropic | ProviderFormat::Converse
-    )
-}
-
-/// Preserve the legacy proxy behavior that fetched remote image URLs before
-/// Bedrock request translation. Lingua still owns source detection and target
-/// serialization; this hook only prepares the normalized universal request
-/// because the network fetch is async and Bedrock targets expect inline media bytes.
-///
-/// This is idempotent: once a remote URL has been replaced by inline base64 data,
-/// subsequent runs leave the image untouched.
-async fn prepare_bedrock_universal_request_with_fetch<F>(
-    request: &mut lingua::UniversalRequest,
-    target_format: ProviderFormat,
-    fetch: F,
-) -> std::result::Result<(), TransformError>
-where
-    F: for<'a> Fn(&'a str) -> FetchMediaFuture<'a> + Send + Sync,
-{
-    if !should_inline_remote_image_urls(target_format) {
-        return Ok(());
-    }
-
-    inline_remote_image_urls_with_fetch(request, fetch)
-        .await
-        .map_err(|err| TransformError::ValidationFailed {
-            target: target_format,
-            reason: err.to_string(),
-        })
-}
-
-async fn inline_remote_image_urls_with_fetch<F>(
-    request: &mut lingua::UniversalRequest,
-    fetch: F,
-) -> Result<()>
-where
-    F: for<'a> Fn(&'a str) -> FetchMediaFuture<'a>,
-{
-    for message in &mut request.messages {
-        let content = match message {
-            Message::System { content }
-            | Message::Developer { content }
-            | Message::User { content } => content,
-            Message::Assistant { .. } | Message::Tool { .. } => continue,
-        };
-
-        let UserContent::Array(parts) = content else {
-            continue;
-        };
-
-        for part in parts {
-            let UserContentPart::Image {
-                image, media_type, ..
-            } = part
-            else {
-                continue;
-            };
-
-            let Some(url) = image.as_str() else {
-                continue;
-            };
-
-            if !is_remote_image_url(url) {
-                continue;
-            }
-
-            let media_block = fetch(url).await?;
-            *image = lingua::serde_json::Value::String(media_block.data);
-            *media_type = Some(media_block.media_type);
-        }
-    }
-
-    Ok(())
-}
-use reqwest_middleware::ClientWithMiddleware;
+use lingua::ProviderFormat;
 
 #[derive(Debug, Clone)]
 pub struct BedrockConfig {
@@ -361,17 +264,6 @@ impl crate::providers::Provider for BedrockProvider {
         vec![ProviderFormat::Converse, ProviderFormat::BedrockAnthropic]
     }
 
-    async fn prepare_universal_request(
-        &self,
-        request: &mut lingua::UniversalRequest,
-        ctx: lingua::RequestPreparationContext,
-    ) -> std::result::Result<(), TransformError> {
-        prepare_bedrock_universal_request_with_fetch(request, ctx.target_format, |url| {
-            Box::pin(fetch_remote_image_as_base64(url))
-        })
-        .await
-    }
-
     async fn complete(
         &self,
         payload: Bytes,
@@ -449,10 +341,6 @@ impl crate::providers::Provider for BedrockProvider {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::catalog::{ModelFlavor, ModelSpec};
-    use async_trait::async_trait;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::Arc;
 
     fn provider() -> BedrockProvider {
         let config = BedrockConfig {
@@ -461,64 +349,6 @@ mod tests {
             timeout: None,
         };
         BedrockProvider::new(config).unwrap()
-    }
-
-    fn bedrock_spec(model: &str, format: ProviderFormat) -> ModelSpec {
-        ModelSpec {
-            model: model.to_string(),
-            format,
-            flavor: ModelFlavor::Chat,
-            display_name: None,
-            parent: None,
-            input_cost_per_mil_tokens: None,
-            output_cost_per_mil_tokens: None,
-            input_cache_read_cost_per_mil_tokens: None,
-            multimodal: None,
-            reasoning: None,
-            max_input_tokens: None,
-            max_output_tokens: None,
-            supports_streaming: true,
-            extra: Default::default(),
-            available_providers: Default::default(),
-        }
-    }
-
-    struct BedrockRequestPreparer<F> {
-        fetch: F,
-    }
-
-    #[async_trait]
-    impl<F> lingua::UniversalRequestPreparer for BedrockRequestPreparer<F>
-    where
-        F: for<'a> Fn(&'a str) -> FetchMediaFuture<'a> + Send + Sync,
-    {
-        async fn prepare_universal_request(
-            &self,
-            request: &mut lingua::UniversalRequest,
-            ctx: lingua::RequestPreparationContext,
-        ) -> std::result::Result<(), TransformError> {
-            prepare_bedrock_universal_request_with_fetch(request, ctx.target_format, &self.fetch)
-                .await
-        }
-    }
-
-    async fn transform_bedrock_request_with_fetch<F>(
-        body: Bytes,
-        spec: &ModelSpec,
-        format: ProviderFormat,
-        fetch: F,
-    ) -> std::result::Result<lingua::TransformResult, TransformError>
-    where
-        F: for<'a> Fn(&'a str) -> FetchMediaFuture<'a> + Send + Sync,
-    {
-        let preparer = BedrockRequestPreparer { fetch };
-        lingua::transform_request_with_universal_preparation(
-            body,
-            format,
-            Some(&spec.model),
-            &preparer,
-        )
-        .await
     }
 
     #[test]
@@ -566,209 +396,5 @@ mod tests {
             }
             other => panic!("expected Error::Auth, got {other:?}"),
         }
-    }
-
-    #[test]
-    fn should_inline_remote_image_urls_matches_bedrock_formats() {
-        assert!(should_inline_remote_image_urls(
-            ProviderFormat::BedrockAnthropic
-        ));
-        assert!(should_inline_remote_image_urls(ProviderFormat::Converse));
-        assert!(!should_inline_remote_image_urls(ProviderFormat::Anthropic));
-        assert!(!should_inline_remote_image_urls(
-            ProviderFormat::ChatCompletions
-        ));
-        assert!(!should_inline_remote_image_urls(ProviderFormat::Responses));
-        assert!(!should_inline_remote_image_urls(ProviderFormat::Google));
-    }
-
-    #[tokio::test]
-    async fn prepare_request_passes_through_same_format_converse_without_fetch() {
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "modelId": "anthropic.claude-3-haiku-20240307-v1:0",
-                "messages": [{
-                    "role": "user",
-                    "content": [{"text": "Hello"}]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let prepared = transform_bedrock_request_with_fetch(
-            body.clone(),
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::Converse,
-            ),
-            ProviderFormat::Converse,
-            |_url| {
-                Box::pin(async {
-                    panic!("fetch should not be called for same-format converse requests");
-                })
-            },
-        )
-        .await
-        .unwrap();
-
-        assert!(prepared.is_passthrough());
-        assert_eq!(prepared.into_bytes(), body);
-    }
-
-    #[tokio::test]
-    async fn prepare_request_inlines_remote_chat_image_for_converse() {
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "model": "claude-sonnet-4-5-20250929",
-                "messages": [{
-                    "role": "user",
-                    "content": [
-                        {"type": "text", "text": "What is this?"},
-                        {"type": "image_url", "image_url": {"url": "https://example.com/image.jpg"}}
-                    ]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let prepared = transform_bedrock_request_with_fetch(
-            body,
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::Converse,
-            ),
-            ProviderFormat::Converse,
-            |_url| {
-                Box::pin(async {
-                    Ok(MediaBlock {
-                        media_type: "image/jpeg".to_string(),
-                        data: "abcd".to_string(),
-                    })
-                })
-            },
-        )
-        .await
-        .unwrap();
-        let value: lingua::serde_json::Value =
-            lingua::serde_json::from_slice(prepared.as_bytes()).unwrap();
-
-        let bytes = value
-            .pointer("/messages/0/content/1/image/source/bytes")
-            .and_then(|v| v.as_str())
-            .unwrap();
-        assert!(!bytes.is_empty());
-        assert_eq!(
-            value
-                .pointer("/messages/0/content/1/image/format")
-                .and_then(|v| v.as_str()),
-            Some("jpeg")
-        );
-    }
-
-    #[tokio::test]
-    async fn prepare_request_inlines_remote_responses_image_for_bedrock_anthropic() {
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "model": "claude-sonnet-4-5-20250929",
-                "input": [{
-                    "role": "user",
-                    "content": [
-                        {"type": "input_text", "text": "What is this?"},
-                        {
-                            "type": "input_image",
-                            "image_url": "https://example.com/image.jpg",
-                            "detail": "auto"
-                        }
-                    ]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let prepared = transform_bedrock_request_with_fetch(
-            body,
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::BedrockAnthropic,
-            ),
-            ProviderFormat::BedrockAnthropic,
-            |_url| {
-                Box::pin(async {
-                    Ok(MediaBlock {
-                        media_type: "image/jpeg".to_string(),
-                        data: "abcd".to_string(),
-                    })
-                })
-            },
-        )
-        .await
-        .unwrap();
-        let value: lingua::serde_json::Value =
-            lingua::serde_json::from_slice(prepared.as_bytes()).unwrap();
-
-        assert_eq!(
-            value.get("anthropic_version").and_then(|v| v.as_str()),
-            Some("bedrock-2023-05-31")
-        );
-        assert_eq!(
-            value
-                .pointer("/messages/0/content/1/source/type")
-                .and_then(|v| v.as_str()),
-            Some("base64")
-        );
-        assert!(value
-            .pointer("/messages/0/content/1/source/data")
-            .and_then(|v| v.as_str())
-            .is_some_and(|v| !v.is_empty()));
-        assert_eq!(value.pointer("/messages/0/content/1/source/url"), None);
-    }
-
-    #[tokio::test]
-    async fn prepare_request_returns_validation_error_when_remote_image_fetch_fails() {
-        let fetch_calls = Arc::new(AtomicUsize::new(0));
-        let body = Bytes::from(
-            lingua::serde_json::to_vec(&lingua::serde_json::json!({
-                "model": "claude-sonnet-4-5-20250929",
-                "messages": [{
-                    "role": "user",
-                    "content": [
-                        {"type": "text", "text": "What is this?"},
-                        {"type": "image_url", "image_url": {"url": "https://example.com/image.jpg"}}
-                    ]
-                }]
-            }))
-            .unwrap(),
-        );
-
-        let err = transform_bedrock_request_with_fetch(
-            body,
-            &bedrock_spec(
-                "anthropic.claude-3-haiku-20240307-v1:0",
-                ProviderFormat::Converse,
-            ),
-            ProviderFormat::Converse,
-            {
-                let fetch_calls = Arc::clone(&fetch_calls);
-                move |url| {
-                    fetch_calls.fetch_add(1, Ordering::SeqCst);
-                    Box::pin(async move {
-                        Err(Error::InvalidRequest(format!(
-                            "failed to fetch image URL {url}: network error"
-                        )))
-                    })
-                }
-            },
-        )
-        .await
-        .expect_err("fetch failure should surface as a validation error");
-
-        assert_eq!(fetch_calls.load(Ordering::SeqCst), 1);
-        assert!(matches!(
-            err,
-            TransformError::ValidationFailed {
-                target: ProviderFormat::Converse,
-                ref reason
-            } if reason.contains("failed to fetch image URL")
-        ));
     }
 }
