+++ crates/lingua/src/processing/mod.rs
@@ -11,7 +11,5 @@ pub use dedup::deduplicate_messages;
 pub use import::{import_and_deduplicate_messages, import_messages_from_spans, Span};
 pub use transform::{
     extract_model, parse_stream_event, response_to_universal, sanitize_payload, transform_request,
-    transform_request_with_universal_preparation, transform_response, transform_stream_chunk,
-    ParsedStreamEvent, RequestPreparationContext, TransformError, TransformResult,
-    UniversalRequestPreparer,
+    transform_response, transform_stream_chunk, ParsedStreamEvent, TransformError, TransformResult,
 };
